Abstracted code for blur outline.

+++ HW3D/BlurOutlinePS.hlsl
@@ -14,7 +14,7 @@ cbuffer Control
 
 float4 main( float2 uv : TexCoord ) : SV_Target
 {
-    uint width, height;
+    float width, height;
     tex.GetDimensions( width, height );
     float dx, dy;
     if ( horizontal )
